Sprint7/customer service fix

+++ dms/Dms.Integration.Infrastructure/Models/DmsPerson.cs
@@ -10,6 +10,6 @@ public class DmsPerson
     public string LastName { get; set; }
     [MaxLength(11)]
     public long CitizenshipNumber { get; set; }
-    public int ClientNumber { get; set; }
+    public long ClientNumber { get; set; }
 }
