+++ workers/Worker.App/Models/SendSmsTemplateRequest.cs
@@ -6,7 +6,7 @@ public class SendSmsTemplateRequest
     public string templateParams { get; set; }
     public string template { get; set; }
     public string title { get; set; }
-    public int customerNo { get; set; }
+    public long customerNo { get; set; }
     public Phone phone { get; set; }
     public Process process { get; set; }
 }
